PoshScriptProcess: Introduced the POSH Script Process
Added a class to represent the Script Process for POSH, this enables non
blocking notifications utilizing QT's Signal and Slot mechanisms to the
full extend. Practically this means the script process can take a long
time to execute while the main window will be responsive. Similarly the
output of the process executing can now send very large amount of data.

Signed-off-by: Sakis Panou <sakis.panou@xilinx.com>